+++ lib/lab_coat.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "benchmark"
+
 require_relative "lab_coat/version"
 require_relative "lab_coat/observation"
 require_relative "lab_coat/result"