+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.2] - Unreleased
+- use `Benchmark` to capture the duration with more details
+- add `to_h` methods to `Result` and `Observation` for convenience
+
 ## [0.1.1] - 2024-04-08
 - add `#slug` method to `Observation`